+++ packages/enhanced/src/lib/container/runtime/FederationRuntimePlugin.ts
@@ -56,7 +56,8 @@ const EmbeddedRuntimePath = require.resolve(
 
 const federationGlobal = getFederationGlobalScope(RuntimeGlobals);
 
-const onceForCompler = new WeakSet<Compiler>();
+const onceForCompiler = new WeakSet<Compiler>();
+const onceForCompilerEntryMap = new WeakMap<Compiler, string>();
 
 class FederationRuntimePlugin {
   options?: moduleFederationPlugin.ModuleFederationPluginOptions;
@@ -150,41 +151,31 @@ class FederationRuntimePlugin {
     ]);
   }
 
-  static getFilePath(
-    compiler: Compiler,
-    options: moduleFederationPlugin.ModuleFederationPluginOptions,
-    bundlerRuntimePath?: string,
-    experiments?: moduleFederationPlugin.ModuleFederationPluginOptions['experiments'],
-  ) {
-    const containerName = options.name;
-    const hash = createHash(
-      `${containerName} ${FederationRuntimePlugin.getTemplate(
-        compiler,
-        options,
-        bundlerRuntimePath,
-        experiments,
-      )}`,
-    );
-    return path.join(TEMP_DIR, `entry.${hash}.js`);
-  }
   getFilePath(compiler: Compiler) {
-    if (this.entryFilePath) {
-      return this.entryFilePath;
-    }
-
     if (!this.options) {
       return '';
     }
 
+    const existedFilePath = onceForCompilerEntryMap.get(compiler);
+
+    if (existedFilePath) {
+      return existedFilePath;
+    }
+
+    let entryFilePath = '';
     if (!this.options?.virtualRuntimeEntry) {
-      this.entryFilePath = FederationRuntimePlugin.getFilePath(
-        compiler,
-        this.options,
-        this.bundlerRuntimePath,
-        this.options.experiments,
+      const containerName = this.options.name;
+      const hash = createHash(
+        `${containerName} ${FederationRuntimePlugin.getTemplate(
+          compiler,
+          this.options,
+          this.bundlerRuntimePath,
+          this.options.experiments,
+        )}`,
       );
+      entryFilePath = path.join(TEMP_DIR, `entry.${hash}.js`);
     } else {
-      this.entryFilePath = `data:text/javascript;charset=utf-8;base64,${pBtoa(
+      entryFilePath = `data:text/javascript;charset=utf-8;base64,${pBtoa(
         FederationRuntimePlugin.getTemplate(
           compiler,
           this.options,
@@ -193,8 +184,12 @@ class FederationRuntimePlugin {
         ),
       )}`;
     }
-    return this.entryFilePath;
+
+    onceForCompilerEntryMap.set(compiler, entryFilePath);
+
+    return entryFilePath;
   }
+
   ensureFile(compiler: Compiler) {
     if (!this.options) {
       return;
@@ -203,7 +198,7 @@ class FederationRuntimePlugin {
     if (this.options?.virtualRuntimeEntry) {
       return;
     }
-    const filePath = this.getFilePath(compiler);
+    const filePath = this.entryFilePath;
     try {
       fs.readFileSync(filePath);
     } catch (err) {
@@ -227,7 +222,7 @@ class FederationRuntimePlugin {
     this.ensureFile(compiler);
 
     this.federationRuntimeDependency = new FederationRuntimeDependency(
-      this.getFilePath(compiler),
+      this.entryFilePath,
     );
     return this.federationRuntimeDependency;
   }
@@ -275,7 +270,7 @@ class FederationRuntimePlugin {
         },
       );
     } else {
-      const entryFilePath = this.getFilePath(compiler);
+      const entryFilePath = this.entryFilePath;
       modifyEntry({
         compiler,
         prependEntry: (entry: Record<string, EntryDescription>) => {
@@ -432,6 +427,8 @@ class FederationRuntimePlugin {
       );
     }
 
+    this.entryFilePath = this.getFilePath(compiler);
+
     if (this.options?.experiments?.federationRuntime === 'hoisted') {
       new EmbedFederationRuntimePlugin().apply(compiler);
 
@@ -451,11 +448,11 @@ class FederationRuntimePlugin {
       ).apply(compiler);
     }
     // dont run multiple times on every apply()
-    if (!onceForCompler.has(compiler)) {
+    if (!onceForCompiler.has(compiler)) {
       this.prependEntry(compiler);
       this.injectRuntime(compiler);
       this.setRuntimeAlias(compiler);
-      onceForCompler.add(compiler);
+      onceForCompiler.add(compiler);
     }
   }
 }