+++ packages/bridge/vue3-bridge/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @module-federation/bridge-vue3
 
+## 0.8.12
+
+### Patch Changes
+
+- @module-federation/runtime@0.8.12
+- @module-federation/sdk@0.8.12
+- @module-federation/bridge-shared@0.8.12
+
 ## 0.8.11
 
 ### Patch Changes