+++ packages/storybook-addon/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @module-federation/storybook-addon
 
+## 4.0.4
+
+### Patch Changes
+
+- Updated dependencies [9062cee]
+  - @module-federation/enhanced@0.8.12
+  - @module-federation/sdk@0.8.12
+  - @module-federation/utilities@3.1.43
+
 ## 4.0.3
 
 ### Patch Changes